requirements.txt
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
pillow
opencv-python
plyfile
huggingface_hub
safetensors
wandb
matplotlib